+++ src/components/layout/NewSidebar.tsx
@@ -1,0 +1,12 @@
+import GlassCard from "@/components/ui/global/GlassContainer";
+
+const Sidebar = () => {
+  return(
+    <aside className="h-full w-75 min-w-75 p-2">
+	    <GlassCard className="h-full w-full">
+		    sidebar
+	    </GlassCard>
+    </aside>
+  )
+}
+export default Sidebar;
